Update Gauge.js - add limitVisibility array and animate option
In config.data.datasets[0] add an array limitVisibility where an element is true if that limit (value) should be visible; if one
does not create an array, all limits are visible as before.

In options, add an animate boolean where true is animate (default) the gauge and false is just change.